requirement.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
paho-mqtt==1.6.1
Pillow==9.2.0
PyJWT==2.5.0
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3